added automatic directory creation

+++ app/echidna_analysis.py
@@ -7,8 +7,12 @@ parent_dir_path = os.path.dirname(os.path.realpath(__file__))
 print(parent_dir_path)
 
 def analyse_echidna(file_name):
-    echidna_dir = parent_dir_path + '/temp/echidna'
-    shutil.rmtree(echidna_dir)
+    temp_dir = join(parent_dir_path, 'temp')
+    echidna_dir = join(temp_dir, 'echidna')
+    if os.path.exists(echidna_dir):
+        shutil.rmtree(echidna_dir)
+    os.mkdir(echidna_dir)
+    
     cmd = ('echidna', file_name, '--corpus-dir', echidna_dir)
     p = subprocess.run(cmd, capture_output=True, text=True)
     fuzzing_log = p.stdout
@@ -16,7 +20,7 @@ def analyse_echidna(file_name):
     print(fuzzing_log, p.stderr)
     print("---")
 
-    with open(parent_dir_path + '/temp/fuzzing_log.pickle', 'wb') as handle:
+    with open(join(temp_dir, 'fuzzing_log.pickle'), 'wb') as handle:
         pickle.dump(fuzzing_log, handle, protocol=pickle.HIGHEST_PROTOCOL)
     
     print(listdir('./example_contracts'))
